Separate Browse filter functionality into new component
